\name{assignVar}
\alias{assignVar}
\title{assign a variable to the environment}
\usage{
  assignVar(var_name, var_value)
}
\arguments{
  \item{var_name}{the name to give to the variable as a
  string}

  \item{var_value}{the value to assign to \code{var_name}}
}
\description{
  assign a variable to the environment
}

