\name{testEnvironment-package}
\alias{testEnvironment-package}
\alias{testEnvironment}
\docType{package}
\title{
Custom Package Environments
}
\description{
These are some simple functions demonstrating how to work with custom environments for storing non-user
visible data in a package.
}
\details{
\tabular{ll}{
Package: \tab testEnvironment\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2013-05-29\cr
License: \tab CC0\cr
}
Only 3 functions

\code{showAll} - lists all of the variables in the environment
\code{showVar} - list a specific variable
\code{assignVar} - add a variable to the environment
}
\author{
Robert M Flight <rflight79@gmail.com>
}
\references{
R Environments, knitcitations package
}

\keyword{ package }
\seealso{
\code{\link[base:environment]{environment}} ~~
}